Broke Triangle, Rectangle, and new WIP cube shape into separate cpp/hpp files

+++ Source/Rasterizer/Mesh.hpp
@@ -2,7 +2,6 @@
 
 #include "Attribute.hpp"
 #include "Util/Vector.hpp"
-#include <cstdint>
 
 namespace Rasterizer
 {
@@ -10,7 +9,7 @@ namespace Rasterizer
 	{
 		TAttribute<FVector3f> Positions = {};
 		TAttribute<FVector4f> Colors = {};
-		uint32_t const* Indices = nullptr;
+		size_t const* Indices = nullptr;
 		size_t Count = 0; // If Indices is valid, contains the number indices in the array. If Indices is not valid, count is the number of vertices.
 	};
 }
